Had to refactor code again due to EVM byte size limit.
Had to refactor Smart Contract Code again, because EVM 24576 bytes limit when deploying the FranklinDecentralizedMarketplace Contract. I wanted to have the FranklinDecentralizedMarketplace Contract be the ONLY Address that could call the public functions of the FranklinDecentralizedMarketplaceMediation Contract via use of Proxy functions, but when I tried to deploy the FranklinDecentralizedMarketplace Contract, I got an error indication stating that the code data had exceeded the 24576 bytes size limit. Now, I no longer have Procy functions on the FranklinDecentralizedMarketplace Contract and the calling entities must now get a "handle" on the FranklinDecentralizedMarketplaceMediatioon Contract deployed by the FranklinDecentralizedMarketplace Contract, and then call the appropriate functions on the "handle" of the FranklinDecentralizedMarketplaceMediation Contract.