Added Events to Emit
Added Events to Emit for the FranklinDecentralizedMarketplace Contract. Limited it to the when someone buys an Item and to annouce a new one as Patrick Galloway suggested I only needed to do, because when I tried to create and emit events for every change to the Smart Contract, I was not able to get a successful deployment due to the infamous "Contract creation initialization returns data with length of more than 24576 bytes. The deployment will likely fails." message.