Code to prevent Mediator from using another Mediator's IPFS Description
Code to prevent Mediator from using another Mediator's IPFS Description. I tried to add a "addressAssociatedWithMediatorDescription" map to the FranklinDecentralizedMarketplaceMediation Contract, but I was not able to deply to due to exceeding the number of bytes limit by EVM (Ethereum Virtual Machine).

I proceeded to create a third Contract - FranklinDecentralizedMarketplaceData - to offload, but I still was not able to deploy the FranklinDecentralizedMarketplaceMediation Contract due to exceeding the number of bytes limit by EVM (Ethereum Virtual Machine).

I got the OK from Patrick Falloway to move on with the previopus version, but I wanted to create a directory with the work I did to make the imprevements so that if time permits, I can go back at a later time to do the imprevements.